+++ http4k-realtime-core/src/main/kotlin/org/http4k/filter/coreExtensions.kt
@@ -1,0 +1,28 @@
+package org.http4k.filter
+
+import org.http4k.core.RequestContext
+import org.http4k.core.Store
+import org.http4k.sse.SseFilter
+import org.http4k.websocket.WsFilter
+
+fun ServerFilters.InitialiseSseRequestContext(contexts: Store<RequestContext>) = SseFilter { next ->
+    {
+        val context = RequestContext()
+        try {
+            next(contexts(context, it))
+        } finally {
+            contexts.remove(context)
+        }
+    }
+}
+
+fun ServerFilters.InitialiseWsRequestContext(contexts: Store<RequestContext>) = WsFilter { next ->
+    {
+        val context = RequestContext()
+        try {
+            next(contexts(context, it))
+        } finally {
+            contexts.remove(context)
+        }
+    }
+}